system/requirements.txt
Django==1.11.6
django-rest-auth==0.9.5
djangorestframework==3.9.4
pytz==2019.2
six==1.12.0
